requirements.txt
flask
groq
twilio
python-dotenv
gunicorn